requirements.txt
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0
pillow>=9.0.0
rich>=12.0.0
numpy>=1.21.0
pandas>=1.4.0
